+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/AuthorizationDetails.kt
@@ -11,16 +11,17 @@ import kotlinx.serialization.json.JsonArray
 @Serializable
 sealed class AuthorizationDetails {
     /**
-     * OID4VCI: The request parameter `authorization_details` defined in Section 2 of (RFC9396) MUST be used to convey the
-     * details about the Credentials the Wallet wants to obtain. This specification introduces a new authorization details
-     * type `openid_credential` and defines the following parameters to be used with this authorization details type.
+     * OID4VCI: The request parameter `authorization_details` defined in Section 2 of (RFC9396) MUST be used to convey
+     * the details about the Credentials the Wallet wants to obtain. This specification introduces a new authorization
+     * details type `openid_credential` and defines the following parameters to be used with this authorization details
+     * type.
      */
     @Serializable
     @SerialName("openid_credential")
     data class OpenIdCredential(
         /**
-         * OID4VC: REQUIRED when [format] parameter is not present. String specifying a unique identifier of the Credential
-         * being described in [IssuerMetadata.supportedCredentialConfigurations].
+         * OID4VCI: REQUIRED when [format] parameter is not present. String specifying a unique identifier of the
+         * Credential being described in [IssuerMetadata.supportedCredentialConfigurations].
          */
         @SerialName("credential_configuration_id")
         val credentialConfigurationId: String? = null,
@@ -36,15 +37,16 @@ sealed class AuthorizationDetails {
 
         /**
          * OID4VCI: ISO mDL: OPTIONAL. This claim contains the type value the Wallet requests authorization for at the
-         * Credential Issuer. It MUST only be present if the [format] claim is present. It MUST not be present otherwise.
+         * Credential Issuer. It MUST only be present if the [format] claim is present. It MUST not be present
+         * otherwise.
          */
         @SerialName("doctype")
         val docType: String? = null,
 
         /**
          * OID4VCI: ISO mDL: OPTIONAL. Object as defined in Appendix A.3.2 excluding the `display` and `value_type`
-         * parameters. The `mandatory` parameter here is used by the Wallet to indicate to the Issuer that it only accepts
-         * Credential(s) issued with those claim(s).
+         * parameters. The `mandatory` parameter here is used by the Wallet to indicate to the Issuer that it only
+         * accepts Credential(s) issued with those claim(s).
          */
         @SerialName("claims")
         val claims: Map<String, Map<String, RequestedCredentialClaimSpecification>>? = null,
@@ -65,21 +67,19 @@ sealed class AuthorizationDetails {
         val sdJwtVcType: String? = null,
 
         /**
-         * Must contain an entry form [IssuerMetadata.authorizationServers].
+         * OID4VCI: If the Credential Issuer metadata contains an [IssuerMetadata.authorizationServers] parameter, the
+         * authorization detail's locations common data field MUST be set to the Credential Issuer Identifier value.
          */
         @SerialName("locations")
         val locations: Set<String>? = null,
 
         /**
-         * OID4VCI: OPTIONAL. Array of strings, each uniquely identifying a Credential that can be issued using the Access
-         * Token returned in this response. Each of these Credentials corresponds to the same entry in the
-         * [IssuerMetadata.supportedCredentialConfigurations] but can contain different claim values or a
-         * different subset of claims within the claims set identified by that Credential type.
-         * This parameter can be used to simplify the Credential Request, as defined in Section 7.2, where the
-         * `credential_identifier` parameter replaces the format parameter and any other Credential format-specific
-         * parameters in the Credential Request. When received, the Wallet MUST use these values together with an Access
-         * Token in subsequent Credential Requests.
+         * OID4VCI: REQUIRED. Array of strings, each uniquely identifying a Credential Dataset that can be issued using
+         * the Access Token returned in this response. Each of these Credential Datasets corresponds to the same
+         * Credential Configuration in the [IssuerMetadata.supportedCredentialConfigurations]. The Wallet MUST use these
+         * identifiers together with an Access Token in subsequent Credential Requests.
          */
+        // TODO is required in OID4VCI!
         @SerialName("credential_identifiers")
         val credentialIdentifiers: Set<String>? = null,
     ) : AuthorizationDetails()