+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/CredentialRequestProofContainer.kt
@@ -1,11 +1,10 @@
 package at.asitplus.openid
 
-import at.asitplus.openid.OpenIdConstants
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class CredentialRequestProof(
+data class CredentialRequestProofContainer(
     /**
      * OID4VCI: e.g. `jwt`, or `ldp_vp`. See [at.asitplus.openid.OpenIdConstants.ProofType].
      */
@@ -17,14 +16,5 @@ data class CredentialRequestProof(
      * a `jwt` claim containing a JWT defined in Section 7.2.1.1.
      */
     @SerialName("jwt")
-    val jwt: String? = null,
-
-    /**
-     * OID4VCI: A CWT (RFC8392) is used as proof of possession. When [proofType] is `cwt`, a proof object MUST include
-     * a `cwt` claim containing a CWT defined in Section 7.2.1.3.
-     *
-     * Removed in OID4VCI Draft 14, kept here for a bit of backwards-compatibility
-     */
-    @SerialName("cwt")
-    val cwt: String? = null,
+    val jwt: Set<String>? = null,
 )