+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/OAuth2AuthorizationServer.kt
@@ -8,8 +8,9 @@ import at.asitplus.openid.OidcUserInfoExtended
  * Used by [CredentialIssuer] to obtain user data when issuing credentials using OID4VCI.
  */
 interface OAuth2AuthorizationServer {
+
     /**
-     * Used in several fields in [IssuerMetadata], to provide endpoint URLs to clients.
+     * Used in several fields in [at.asitplus.openid.IssuerMetadata], to provide endpoint URLs to clients.
      */
     val publicContext: String
 
@@ -20,13 +21,22 @@ interface OAuth2AuthorizationServer {
     suspend fun providePreAuthorizedCode(): String?
 
     /**
-     * Get the [OidcUserInfoExtended] (holding [OidcUserInfo]) associated with the [accessToken],
+     * Get the [OidcUserInfoExtended] (holding [at.asitplus.openid.OidcUserInfo]) associated with the [accessToken],
      * that was created before at the Authorization Server.
      */
     suspend fun getUserInfo(accessToken: String): KmmResult<OidcUserInfoExtended>
 
-    // TODO How is this supposed to happen when using an external Authorization Server?
-    suspend fun verifyAndRemoveClientNonce(nonce: String): Boolean
+    /**
+     * Whether this authorization server includes [at.asitplus.openid.TokenResponseParameters.clientNonce] it its
+     * token response, i.e. whether the [CredentialIssuer] needs to verify it using [verifyClientNonce].
+     */
+    val supportsClientNonce: Boolean
+
+    /**
+     * Called by [CredentialIssuer] to verify that nonces contained in proof-of-possession statements from clients
+     * are indeed valid.
+     */
+    suspend fun verifyClientNonce(nonce: String): Boolean
 
     /**
      * Provide necessary [OAuth2AuthorizationServerMetadata] JSON for a client to be able to authenticate